Add single_key flag, to reuse the private key
This option is useful when testing with tens of thousands of devices;
the key generation requires intensive computation, which we can avoid
reusing the same key pair for all the devices. As the mac address is
part of the identity attributes of the devices, the backend will
recognize them as different devices.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>